Fix: First use of ServletModule must be within global scope
Every ServletModule instance tries to install InternalServletModule.
The trick is used, because Guice installs modules only the first time
and ignores any subsequent execution of install method with the same
module (by comparing them using equals() method). We need to make sure
InternalServletModule is installed from the top level module,
because when it is installed from some private module it doesn't have
an access to all registered servlets and filters.

Change-Id: Ifade84ecd51f27839fcdf34976263621c7034ba5
Reviewed-on: https://pirx.atteo.com/gerrit/227
Tested-by: Mr. Jenkins
Reviewed-by: Tomek Piotrowski <tomek.piotrowski@atteo.com>