Blueprints API Service.
Blueprints is a collection of interfaces, implementations, ouplementations, and test suites for the property graph data model.
Blueprints is analogous to the JDBC, but for graph databases.
As such, it provides a common set of interfaces to allow developers to plug-and-play their graph database backend.

This commit contains Blueprints service that acts as a base service for any graph DB implementations that support
Blueprints API (such as Neo4j).

This commit also contains modifications of Neo4j service to support Blueprints.

Unit tests on Blueprints level are included.